candidates/requirements.txt
amqp==5.0.9
arabic-reshaper==2.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
billiard==3.6.4.0
black==22.1.0
bleach==4.1.0
blis==0.7.7
Brotli==1.0.9
cachetools==5.0.0
cairocffi==1.3.0
CairoSVG==2.5.2
catalogue==2.0.7
celery==5.2.3
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.0
cssselect2==0.4.1
cymem==2.0.6
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Django==3.1.4
django-celery-results==2.2.0
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-crontab==0.7.1
django-dbbackup==3.3.0
django-filter==2.4.0
django-js-asset==1.2.2
django-rest-authtoken==2.1.4
django-rest-swagger==2.2.0
django-weasyprint==2.1.0
djangorestframework==3.13.1
djangorestframework-csv==2.1.0
docx2txt==0.8
drf-api-logger==1.0.9
elastic-apm==6.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
entrypoints==0.4
executing==0.8.2
flower==1.0.0
fonttools==4.29.0
future==0.18.2
google-api-core==2.4.0
google-api-python-client==2.36.0
google-auth==2.5.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.54.0
gunicorn==20.1.0
html2docx==1.3.1
html5lib==1.1
httplib2==0.20.2
humanize==3.13.1
idna==2.10
ipykernel==6.9.1
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kombu==5.2.3
langcodes==3.3.0
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
murmurhash==1.0.7
mypy-extensions==0.4.3
mysqlclient==2.1.0
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
nltk==3.7
notebook==6.4.8
numpy==1.22.1
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==21.3
pandas==1.2.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pathy==0.6.1
pdfminer.six==20220319
pdfplumber==0.6.1
pexpect==4.8.0
phonenumbers==8.12.47
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.1
preshed==3.0.6
prometheus-client==0.13.1
prompt-toolkit==3.0.24
protobuf==3.19.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
pydyf==0.1.2
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==3.0.7
PyPDF2==1.26.0
pyphen==0.12.0
pyrsistent==0.18.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.3
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
redis==3.5.3
regex==2022.4.24
reportlab==3.6.6
requests==2.25.1
requests-oauthlib==1.3.0
resume-parser==0.8.4
rsa==4.8
Send2Trash==1.8.0
sentry-sdk==1.5.4
simplejson==3.17.6
six==1.16.0
smart-open==5.2.1
spacy==3.2.4
spacy-legacy==3.0.9
spacy-loggers==1.0.2
sqlparse==0.4.2
srsly==2.4.3
stack-data==0.2.0
stemming==1.0.1
terminado==0.13.1
testpath==0.6.0
thinc==8.0.15
tika==1.24
tinycss2==1.1.1
tomli==2.0.1
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typer==0.4.1
typing_extensions==4.2.0
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
Wand==0.6.7
wasabi==0.9.1
wcwidth==0.2.5
weasyprint==54.0
webencodings==0.5.1
widgetsnbextension==3.5.2
xhtml2pdf==0.2.5
zoomus==1.1.3
zopfli==0.1.9
